test: create chat mutation unit tests
