+++ lib/resolvers/direct_chat_mutations/createDirectChat.js
@@ -43,7 +43,7 @@ module.exports = async (parent, args, context) => {
     }
     usersInChat.push(user);
   }
-  user.password = null
+
   let directChat = new DirectChat({
     creator: user,
     users: usersInChat,
